Add files via upload
As a note: The FOLDER name in which your INITIAL folder right after the folder holding all your mods is the name that the launcher will pick. If you suck at naming folders, I can't help ya. I may add functionality later for it to identify which mod is which using some kind of property or asset inside the FNF mod folder, but for the time being, make sure your folders are named by the mods.